Trying out lazygit-confirm plugin
This is to keep me from committing changes with lazygit when I have
(forgotten) unsaved changes in the buffer. It's a little broken at the
moment because the confirmation dialog doesn't show, which appears to be
a neovim bug that was fixed a couple weeks ago (and so hopefully will go
away when I update neovim). See here:

https://github.com/folke/noice.nvim/issues/1036