lazygit-confirm configs (saveall)
Not sure why I can't update neovim (yet) to fix the confirm dialog bug,
but in the meantime the confirm buttons will still work -- I just have
to remember that is why things seem to hang if I run lazygit when I have
unsaved buffers.

In this commit I added the `show_saveall_noconfirm = true` option, which
allows me to hit "a" to save all open buffers and continue to lazygit.